rename: 구조 및 컴포넌트, 파일 이름 세분화 및 최적화
