+++ src/app/error.tsx
@@ -1,8 +1,8 @@
 // app/error.tsx
 'use client';
-import FallbackUI from '@/components/organisms/FallbackUI';
+import ErrorFallback from '@/components/templates/ErrorFallback';
 import { ErrorPageProps } from '@/types/props/errorProps';
 
 export default function Error({ error }: ErrorPageProps) {
-  return <FallbackUI error={error} />;
+  return <ErrorFallback error={error} />;
 }
