+++ src/app/page.tsx
@@ -1,6 +1,6 @@
 export const dynamic = 'force-static';
 
-import HomeTemplate from '@/components/templates/HomeTemplate';
+import HomeTemplate from '@/components/templates/Home';
 import { getPostList } from '@/services/post/post';
 
 export default async function Home() {
