+++ src/components/boundaries/LocalErrorBoundary.tsx
@@ -1,13 +1,13 @@
 'use client';
 
 import { ErrorBoundary as ReactErrorBoundary } from 'react-error-boundary';
-import FallbackUI from '../organisms/FallbackUI';
+import ErrorFallback from '../templates/ErrorFallback';
 
 export default function LocalErrorBoundary({ children }: { children: React.ReactNode }) {
   return (
     <ReactErrorBoundary
       fallbackRender={({ error, resetErrorBoundary }) => (
-        <FallbackUI error={error} onRetry={resetErrorBoundary} />
+        <ErrorFallback error={error} onRetry={resetErrorBoundary} />
       )}
       onError={(error, info) => {
         console.error('Local Error:', error);
