+++ src/components/molecules/post/PostContent.tsx
