+++ src/components/organisms/layout/Header.tsx
@@ -1,14 +1,13 @@
-// components/organisms/Header.tsx
 'use client';
 
-import ThemeToggle from '@/components/atoms/ThemeToggle';
-import Logo from '../atoms/Logo';
+import Logo from '../../atoms/Logo/Logo';
+import ThemeDropdown from '../../molecules/dropdown/ThemeDropdown/ThemeDropdown';
 
 export default function Header() {
   return (
     <header className="flex items-center justify-between w-full z-50 h-18 p-content">
       <Logo />
-      <ThemeToggle />
+      <ThemeDropdown />
     </header>
   );
 }
