+++ src/components/templates/NotFound.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-export default function PageNotFound() {
+export default function NotFound() {
   return (
     <div className="flex flex-col items-center justify-center min-h-screen text-center p-6">
       <h1 className="text-4xl font-bold mb-4">404</h1>
